Bugfix/kevin

+++ Proyecto-REMI-WebApi/Proyecto-REMI-WebApi/Models/detallesPedido.cs
@@ -10,17 +10,18 @@ namespace Proyecto_REMI_WebApi.Models;
 [Table("detallesPedido")]
 public partial class detallesPedido
 {
+
     [Key]
     public int codigoPedido { get; set; }
 
     [Key]
     public int codigoProducto { get; set; }
 
-    public double cantidadProducto { get; set; }
+    public int cantidadProducto { get; set; }
 
-    public double valorProducto { get; set; }
+    public decimal valorProducto { get; set; }
 
-    public double totalPagoProducto { get; set; }
+    public decimal totalPagoProducto { get; set; }
 
     [ForeignKey("codigoPedido")]
     [InverseProperty("detallesPedidos")]
